refactor: getTodayDate formater 함수 utils로 이동
